Only allow track Player death once
There was a weird issue where a Player could alert the PlayerManager of
it's death multiple times resulting in frequent negative _playerCount in
PlayerManager. I created an public property IsDead on the Player that
gets set to true when the Player collides with a wall the first time. I
then check if isDead is true before handling death logic. This prevents
multiple collisions from a single Player counting as multiple deaths.

+++ game/managers/PlayerManager.cs
@@ -32,6 +32,11 @@ public partial class PlayerManager : Node
 	// ! Tracking player numbers this way is highly inaccurate
 	public void KillPlayer(Player playerToKill)
 	{
+		// Ensures a Player's death is not counted multiple times
+		if (playerToKill.IsDead)
+			return;
+
+		playerToKill.IsDead = true;
 		playerToKill.Kill();
 		_playerCount--;
 		GD.Print("Updated player count: " + _playerCount);
